add middlewares and controllers
